updates to npm, new editor component
