+++ src/components/style.css
@@ -1,4 +1,4 @@
-* {
+/* * {
     outline: solid 1px rgba(255, 0, 170, 0.507);
     margin: 0;
     padding: 0;
@@ -59,4 +59,4 @@
 
   #X {
     background-color: gray;
-  }
+  } */
